refactor IndexReader (#1068)
move IndexReader from pyserini.index to pyserini.index.lucene
